Move cipher and shanda to core crypt mod

+++ sql/sessions.sql
@@ -2,7 +2,7 @@ CREATE TABLE sessions (
     id integer NOT NULL PRIMARY KEY,
     account_id integer NOT NULL,
     character_id integer NOT NULL,
-    world_id integer NOT NULL,
+    world_id smallint NOT NULL,
     channel_id smallint NOT NULL,
 );
